requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Claude API
anthropic==0.18.0

# Trading & Market Data
alpaca-py==0.13.0
yfinance==0.2.33
pandas==2.1.4
numpy==1.26.2
ta==0.11.0  # Technical analysis
pandas-ta==0.3.14b0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# API & Web
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Data Sources
polygon-api-client==1.13.0
alpha-vantage==2.3.1
newsapi-python==0.2.7

# Scheduling & Async
asyncio==3.4.3
aiocron==1.8

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
faker==20.1.0

# Utilities
pytz==2023.3
python-dateutil==2.8.2
click==8.1.7  # CLI interface

# Data validation
marshmallow==3.20.1
jsonschema==4.20.0

# Risk Management
pyfolio==0.9.2  # Portfolio analysis
quantstats==0.0.62  # Performance metrics